+++ packages/language-service/index.ts
@@ -1,9 +1,9 @@
 export * from '@volar/language-core';
-export * from './lib/baseLanguageService';
+export * from './lib/languageService';
 export * from './lib/documents';
 export { mergeWorkspaceEdits } from './lib/languageFeatures/rename';
 export * from './lib/types';
-export * as transformer from './lib/utils/transform';
+export * from './lib/utils/transform';
 
 import type * as vscode from 'vscode-languageserver-protocol';
